Prerelease

+++ Sketchbook/libraries/LovyanGFX/src/lgfx/v1/panel/Panel_RA8875.hpp
@@ -63,7 +63,7 @@ namespace lgfx
     void setInvert(bool invert) override {}; // Not yet implemented.
     void setSleep(bool flg) override {}; // Not yet implemented.
     void setPowerSave(bool flg) override {}; // Not yet implemented.
-    uint32_t readCommand(uint_fast8_t cmd, uint_fast8_t index, uint_fast8_t len) override { return 0; }; // Not yet implemented.
+    uint32_t readCommand(uint_fast16_t cmd, uint_fast8_t index, uint_fast8_t len) override { return 0; }; // Not yet implemented.
     uint32_t readData(uint_fast8_t index, uint_fast8_t len)  override { return 0; }; // Not yet implemented.
 
   protected:
@@ -86,7 +86,7 @@ namespace lgfx
 
     void _set_write_pos(uint_fast16_t x, uint_fast16_t y);
     bool _wait_busy( uint32_t timeout = 1000);
-    void _write_reg(uint8_t reg, uint8_t data);
+    void _write_reg(uint_fast16_t reg, uint_fast16_t data);
     void _write_reg_0x51(uint8_t reg, uint8_t data);
     void _start_memorywrite(void);
   };
